Move headers (and cpp files) to a qtkeychain subdir
... and use target_include_directories() so that one can
include <qtkeychain/keychain.h> rather than <keychain.h>
when using this repository as a git submodule.
