+++ qtkeychain/gnomekeyring.cpp
