+++ qtkeychain/gnomekeyring_p.h
