+++ qtkeychain/keychain_haiku.cpp
