+++ testclient.cpp
@@ -16,7 +16,7 @@
 
 #include <QStringList>
 
-#include "keychain.h"
+#include "qtkeychain/keychain.h"
 #include <iostream>
 
 using namespace QKeychain;
